fix(cooldown): use cooldown on command instead of compose

+++ packages/backend/src/modules/app-stores/repos.helpers.ts
@@ -148,7 +148,9 @@ export class ReposHelpers {
       this.logger.debug(`Pulled repo ${repoUrl} to ${repoPath}`);
       return { success: true, message: '' };
     } catch (_) {
-      await this.deleteRepo(slug);
+      if (this.configuration.get('__prod__')) {
+        await this.deleteRepo(slug);
+      }
       return this.cloneRepo(repoUrl, slug);
     }
   }
